Decouple InfoPanel and EntityInfo
